Split TWI off into its own file
